+++ src/models/export.cr
@@ -7,14 +7,14 @@ class Export < BaseModel
     field year : Int32
     field month : Int32?
     field cpc : String?
-    field quantity : Int32?
+    field quantity : Int64?
 
-    field mass_gross_kg : Int32?
-    field mass_net_kg : Int32
-    field fob_etb_cents : Int32
-    field fob_usd_cents : Int32
-    field tax_etb_cents : Int32?
-    field tax_usd_cents : Int32?
+    field mass_gross_kg : Int64?
+    field mass_net_kg : Int64
+    field fob_etb_cents : Int64
+    field fob_usd_cents : Int64
+    field tax_etb_cents : Int64?
+    field tax_usd_cents : Int64?
 
     field unique_hash : String
   end
